server/requirements.txt
Django==1.8.3
django-filter==0.10.0
django-guardian==1.3
djangorestframework==3.2.0
